Tidy up query performance page
